+++ crates/minecraft_packets/src/configuration/server_bound_known_packs_packet.rs
@@ -3,5 +3,20 @@ use minecraft_protocol::prelude::*;
 
 #[derive(PacketIn)]
 pub struct ServerBoundKnownPacksPacket {
-    pub known_packs: LengthPaddedVec<KnownPack>,
+    known_packs: LengthPaddedVec<KnownPack>,
+}
+
+impl ServerBoundKnownPacksPacket {
+    pub fn new(known_packs: Vec<KnownPack>) -> Self {
+        Self {
+            known_packs: LengthPaddedVec::new(known_packs),
+        }
+    }
+
+    pub fn has_minecraft_core(&self) -> bool {
+        self.known_packs
+            .inner()
+            .iter()
+            .any(KnownPack::is_minecraft_core)
+    }
 }