+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - The plugin doesn't shutdown when shutting down the proxy
-- Do not send registry data starting 1.21.5, this should fix issues with PacketEvents some users were having
+- PicoLimbo should wait for client known packs before sending registry data (starting 1.20.5)
+- Registry data can be omitted if it is known by the client (starting 1.21.5)
 
 ## [1.12.2+mc26.1.2] - 2026-04-12